feat: 完成 Act 模块重构并完善 Observe 模块
Act 模块：
- 新增 history.rs 实现撤销/重做系统
- 新增 action_state.rs 管理动作状态
- 新增 validation.rs 验证动作配置
- 新增 zmq_state.rs 和 zmq_pull.rs 重构 ZMQ 连接
- 重构 config.rs 精简代码

Observe 模块：
- 新增 zmq_pub.rs 实现 ZMQ PUB 发布

代码清理：
- 删除 state.rs 和旧的 zmq_pub/zmq_pull 模块
- 完善 lib.rs 模块化拆分

前端：
- 重构 ObservePage 子组件（CaptureTabPanel、CropTabPanel、TransportTabPanel）
- 添加 ESLint flat config 配置
- 完善 Act 模块编辑器组件

+++ frontend/src-tauri/src/observe/capture.rs
@@ -87,8 +87,8 @@ impl GraphicsCaptureApiHandler for WgcFrameHandler {
         }
 
         // 2. 如有需要则缩放
-        let scaled = if self.data.config.capture.target_width != width as u32
-            || self.data.config.capture.target_height != height as u32
+        let scaled = if self.data.config.capture.target_width != width
+            || self.data.config.capture.target_height != height
         {
             let processor = ImageProcessor {};
             processor.scale(
